checkpoint; preliminary extrapolation working
