+++ src/mplfinance/_arg_validators.py
@@ -5,8 +5,6 @@ import datetime
 from   mplfinance._helpers import _list_of_dict
 import matplotlib as mpl
 import warnings
-from   mplfinance._utils import _date_to_mdate
-from   mplfinance._utils import _date_to_iloc_extrapolate
 
 def _check_and_prepare_data(data, config):
     '''
@@ -77,28 +75,6 @@ def _check_and_prepare_data(data, config):
             raise ValueError('Data for column "'+str(col)+'" must be ALL float or int.')
 
     return dates, opens, highs, lows, closes, volumes
-
-def _check_and_convert_xlim_configuration(data, config):
-    '''
-    Check, if user entered `xlim` kwarg, if user entered dates
-    then we may need to convert them to iloc or matplotlib dates.
-    '''
-    if config['xlim'] is None:
-        return None
-
-    xlim = config['xlim']
-
-    if not _xlim_validator(xlim):
-        raise ValueError('Bad xlim configuration #1')
-
-    if all([_is_date_like(dt) for dt in xlim]):
-        if config['show_nontrading']:
-            xlim = [ _date_to_mdate(dt) for dt in xlim]
-        else
-            xlim = [ _date_to_iloc_extrapolate(data.index.to_series(),dt) for dt in xlim]
-        
-    return xlim
-
 
 def _get_valid_plot_types(plottype=None):
